move: 불필요한 폴더(main_page) 제거
