+++ src/main.jsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import GlobalStyles from './styles/GlobalStyles.jsx'
-import MainPage from './main_page/MainPage.jsx'
 import { Provider } from 'react-redux'
+import MainPage from './MainPage.jsx'
 import Store from './store/Store.js'
+import GlobalStyles from './styles/GlobalStyles.jsx'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
 	<React.StrictMode>
